# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assessment API",
        "contact": {
            "email": "ajiferukeolatommy@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/todos": {
            "get": {
                "tags": [
                    "HR/Director Operations"
                ],
                "summary": "Get employees data",
                "description": "An authenticated user retrieving employees data using his/her ID",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "Bearer-Token",
                        "in": "header",
                        "description": "Set logged in user token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Employees[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "HR/Director Operations"
                ],
                "summary": "Update employee data",
                "description": "An authenticated user update an employee data using his/her ID",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "Bearer-Token",
                        "in": "header",
                        "description": "Set logged in user token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullName",
                                    "professionalQualification",
                                    "educationDegree",
                                    "dob",
                                    "phoneNumber",
                                    "nextOfKin",
                                    "maritalStatus",
                                    "homeAddress"
                                ],
                                "properties": {
                                    "fullName": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Mr. Ajagbe Daniel Edited"
                                    },
                                    "professionalQualification": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Professor"
                                    },
                                    "educationDegree": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Masters"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1998/07/03"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "08023456782"
                                    },
                                    "nextOfKin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Ajagbe Tolani"
                                    },
                                    "maritalStatus": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Single"
                                    },
                                    "homeAddress": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Baruwa Street"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Employee{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Oops, the todo could not be updated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "HR/Director Operations"
                ],
                "summary": "Add employee data",
                "description": "An authenticated user add new employee data using his/her ID",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "Bearer-Token",
                        "in": "header",
                        "description": "Set logged in user token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullName",
                                    "professionalQualification",
                                    "educationDegree",
                                    "dob",
                                    "phoneNumber",
                                    "nextOfKin",
                                    "maritalStatus",
                                    "homeAddress"
                                ],
                                "properties": {
                                    "fullName": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Mr. Ajagbe Daniel Edited"
                                    },
                                    "professionalQualification": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Professor"
                                    },
                                    "educationDegree": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Masters"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1998/07/03"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "08023456782"
                                    },
                                    "nextOfKin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Ajagbe Tolani"
                                    },
                                    "maritalStatus": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Single"
                                    },
                                    "homeAddress": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Baruwa Street"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Employee{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Oops, the todo could not be updated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}